Formatting returned documents.
Before this PR, documents are just json serialized
by tantivy.

This is bad because:
- object fields are not constructed,
- cardinality is ignored, and all fields are wrapped in arrays,
- the dynamically mapped fields are added in `_dynamic`,

This PR adds a method to the `DocMapper` in
order to format JSON Object for documents.

This formatting happens at the root level, in order
to use the same DocMapper (output format) for all hits.

In addition, the _source field used to be stored in
text field. After this change, the _source field is stored in a
JSON field, and returned as a JSON Object (it used to be a JSON
serialized string within the json document).

Closes #1325
Closes #1293